Week tests made and organized, still haven't figured out plants lol
