+++ src/main/java/com/zipcodewilmington/froilansfarm/Plants/CornStalk.java
@@ -1,7 +1,21 @@
 package com.zipcodewilmington.froilansfarm.Plants;
 
+import com.zipcodewilmington.froilansfarm.Interfaces.Edible;
+
 public class CornStalk extends Crop {
 
+    @Override
+    public Edible yield() {
+        if (fertilized) {
+            return new EarCorn();
+        } else {
+            return null;
+        }
+    }
 
+    @Override
+    public boolean isFertilized() {
+        return super.isFertilized();
+    }
 
 }
