+++ src/test/java/com/zipcodewilmington/froilansfarm/OutdatedTests/AnimalTest.java
@@ -1,4 +1,4 @@
-package com.zipcodewilmington.froilansfarm;
+package com.zipcodewilmington.froilansfarm.OutdatedTests;
 
 import com.zipcodewilmington.froilansfarm.Animals.Horse;
 import com.zipcodewilmington.froilansfarm.People.Farmer;
