+++ src/test/java/com/zipcodewilmington/froilansfarm/WeekdayTests/MondayTest.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.WeekdayTests;
+
+public class MondayTest {
+}
