+++ src/test/java/com/zipcodewilmington/froilansfarm/WeekdayTests/SundayTest.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.WeekdayTests;
+
+public class SundayTest {
+}
